# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mirage private merchant gateway API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PaymentRecord {
  id: number;
  amount: number;
  currency: string;
  facadeAddress: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  merchantId?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  amount: number;
  currency: string;
  facadeAddress: string;
  sessionId?: string;
  txHash?: string;
  merchantId?: string;
}

export interface PaymentStats {
  totalUSDC: number;
  totalPayments: number;
  recent: PaymentRecord[];
}

export interface SessionRecord {
  id: string;
  facadeAddress: string;
  label: string;
  expiryMinutes: number;
  /** @nullable */
  amount?: number | null;
  currency: string;
  /** @nullable */
  merchantId?: string | null;
  status: string;
  createdAt: string;
  expiresAt: string;
  checkoutUrl?: string;
}

export interface SessionInput {
  label: string;
  expiryMinutes: number;
  amount?: number;
  currency?: string;
  merchantId?: string;
}

export interface ApiKeyRecord {
  id: number;
  merchantId: string;
  keyPrefix: string;
  label: string;
  createdAt: string;
  /** @nullable */
  lastUsedAt?: string | null;
}

export interface ApiKeyInput {
  merchantId: string;
  label: string;
}

export interface ApiKeyCreated {
  id: number;
  merchantId: string;
  keyPrefix: string;
  label: string;
  createdAt: string;
  /** @nullable */
  lastUsedAt?: string | null;
  rawKey: string;
}

export type ListPaymentsParams = {
merchantId?: string;
};

export type GetPaymentStatsParams = {
merchantId?: string;
};

export type ListSessionsParams = {
merchantId?: string;
};

export type ListApiKeysParams = {
merchantId: string;
};