# lib/api-zod/src/generated/types/apiKeyRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mirage private merchant gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKeyRecord {
  id: number;
  merchantId: string;
  keyPrefix: string;
  label: string;
  createdAt: string;
  /** @nullable */
  lastUsedAt?: string | null;
}