# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mirage private merchant gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInput {
  amount: number;
  currency: string;
  facadeAddress: string;
  sessionId?: string;
  txHash?: string;
  merchantId?: string;
}