# lib/api-zod/src/generated/types/paymentRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mirage private merchant gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentRecord {
  id: number;
  amount: number;
  currency: string;
  facadeAddress: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  merchantId?: string | null;
  createdAt: string;
}