# lib/api-zod/src/generated/types/paymentStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mirage private merchant gateway API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRecord } from './paymentRecord';

export interface PaymentStats {
  totalUSDC: number;
  totalPayments: number;
  recent: PaymentRecord[];
}