# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mirage private merchant gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  label: string;
  expiryMinutes: number;
  amount?: number;
  currency?: string;
  merchantId?: string;
}