# lib/api-zod/src/generated/types/sessionRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mirage private merchant gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionRecord {
  id: string;
  facadeAddress: string;
  label: string;
  expiryMinutes: number;
  /** @nullable */
  amount?: number | null;
  currency: string;
  /** @nullable */
  merchantId?: string | null;
  status: string;
  createdAt: string;
  expiresAt: string;
  checkoutUrl?: string;
}